Refactor Dht11 class and improve detection method
The name of the Dht11 class was corrected from Dth11 to Dht11. The detection method was significantly reworked, replacing several methods and literals with new functions and constants to better handle data readouts from a device. The AbstractOneWireDevice class was also simplified and updated to complement these changes in the Dht11 class.