Rename package path pi to rpi
The code changes consist of renaming the package path from "pi" to "rpi" across various java files, and updating dependencies and repository details in pom.xml. The "pi" package and its sub-packages are renamed, and all the import statements are updated accordingly. This name change is more reflective of the project's target platform, Raspberry Pi.