Update DHT device handling and optimize data timing
The DHT device handling was updated and revised to be more efficient. A key change is that I/O operations with pins of 100us or less can't be guaranteed due to Java execution efficiency issues, preventing the use of the current solution to read DHT data. Another modification optimized the send data time from 20ms to 18ms, and also simplified the data detection and processing methods. The README was amended to reflect these changes.